# schemas/minos_backend_openapi.json
{
  "components": {
    "securitySchemes": {
      "AccountBearer": {
        "bearerFormat": "JWT",
        "description": "Bearer token used by account_api endpoints.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated from minos-backend's formal route inventory. Operation bodies stay intentionally schematic in this pass; the document captures mounted paths, methods, auth surfaces, and websocket upgrade points.",
    "title": "Minos Backend HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/hosts": {
      "get": {
        "description": "Generated from the mounted route inventory for the `account_api` surface with `account_bearer` auth semantics.",
        "operationId": "get_v1_hosts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": true
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required or rejected"
          }
        },
        "security": [
          {
            "AccountBearer": []
          }
        ],
        "summary": "GET /v1/hosts",
        "tags": [
          "account_api"
        ],
        "x-minos-auth": "account_bearer",
        "x-minos-probe-path": "/v1/hosts",
        "x-minos-surface": "account_api"
      }
    },
    "/v1/hosts/link": {
      "post": {
        "description": "Generated from the mounted route inventory for the `account_api` surface with `account_bearer` auth semantics.",
        "operationId": "post_v1_hosts_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": true
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": true
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required or rejected"
          }
        },
        "security": [
          {
            "AccountBearer": []
          }
        ],
        "summary": "POST /v1/hosts/link",
        "tags": [
          "account_api"
        ],
        "x-minos-auth": "account_bearer",
        "x-minos-probe-path": "/v1/hosts/link",
        "x-minos-surface": "account_api"
      }
    },
    "/v1/hosts/unlink": {
      "post": {
        "description": "Generated from the mounted route inventory for the `account_api` surface with `account_bearer` auth semantics.",
        "operationId": "post_v1_hosts_unlink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": true
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": true
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required or rejected"
          }
        },
        "security": [
          {
            "AccountBearer": []
          }
        ],
        "summary": "POST /v1/hosts/unlink",
        "tags": [
          "account_api"
        ],
        "x-minos-auth": "account_bearer",
        "x-minos-probe-path": "/v1/hosts/unlink",
        "x-minos-surface": "account_api"
      }
    },
...
    "/v1/realtime/ws-ticket": {
      "post": {
        "description": "Generated from the mounted route inventory for the `account_api` surface with `account_bearer` auth semantics.",
        "operationId": "post_v1_realtime_ws_ticket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": true
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": true
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Authentication required or rejected"
          }
        },
        "security": [
          {
            "AccountBearer": []
          }
        ],
        "summary": "POST /v1/realtime/ws-ticket",
        "tags": [
          "account_api"
        ],
        "x-minos-auth": "account_bearer",
        "x-minos-probe-path": "/v1/realtime/ws-ticket",
        "x-minos-surface": "account_api"
      }
    },
    "/ws/client": {
      "get": {
        "description": "WebSocket upgrade endpoint for the `client_gateway` surface. The frame contract is emitted separately in schemas/minos_backend_ws_schema.json.",
        "operationId": "get_ws_client",
        "responses": {
          "101": {
            "description": "WebSocket upgrade accepted"
          },
          "401": {
            "description": "Upgrade rejected before websocket activation"
          }
        },
        "summary": "GET /ws/client",
        "tags": [
          "client_gateway"
        ],
        "x-minos-auth": "realtime_ticket",
        "x-minos-probe-path": "/ws/client",
        "x-minos-surface": "client_gateway",
        "x-minos-websocket": true,
        "x-minos-websocket-schema-ref": "#/schemas/Envelope"
      }
    },
    "/ws/host": {
      "get": {
        "description": "WebSocket upgrade endpoint for the `host_gateway` surface. The frame contract is emitted separately in schemas/minos_backend_ws_schema.json.",
        "operationId": "get_ws_host",
        "responses": {
          "101": {
            "description": "WebSocket upgrade accepted"
          },
          "401": {
            "description": "Upgrade rejected before websocket activation"
          }
        },
        "summary": "GET /ws/host",
        "tags": [
          "host_gateway"
        ],
        "x-minos-auth": "realtime_ticket",
        "x-minos-probe-path": "/ws/host",
        "x-minos-surface": "host_gateway",
        "x-minos-websocket": true,
        "x-minos-websocket-schema-ref": "#/schemas/Envelope"
      }
    }
  },
...
}